fix broken import after utility rename
