+++ streamwebs_frontend/streamwebs/tests/views/test_camerapoint_view.py
@@ -1,7 +1,7 @@
 from django.test import Client, TestCase
 from django.core.urlresolvers import reverse
 from streamwebs.models import Site, PhotoPointImage, PhotoPoint, CameraPoint
-from streamwebs.util.temp_img import get_temporary_image
+from streamwebs.util.create_dummy_files import get_temporary_image
 
 
 class ViewCameraPointTestCase(TestCase):
